requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
based==0.7.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
deptry==0.23.0
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
greenlet==3.2.2
idna==3.10
mock==5.2.0
packaging==25.0
pillow==11.2.1
pip-autoremove==0.10.0
pipdeptree==2.26.1
projects==1.0.9
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pyotp==2.9.0
pyproject-toml==0.1.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requirements-parser==0.11.0
setuptools==80.4.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
termcolor==3.1.0
types-setuptools==80.3.0.20250505
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
urwid==3.0.2
wcwidth==0.2.13
wheel==0.45.1
channels==4.2.2
whitenoise==6.9.0
cloudinary==1.44.0
python-dotenv==1.1.0